refactor(tests): C9-10 split test_server_init.py into focused modules
Split the 1,336-line test_server_init.py into three files:
- test_server_tool_registration.py: tool registration, config-driven behavior, schema contracts
- test_server_version_telemetry.py: plugin metadata, version info, lazy init, telemetry
- test_server_init.py: kitchen gate access, visibility, subset management

Create tests/server/conftest.py with shared kitchen_enabled fixture.
Apply kitchen_enabled fixture to TestToolRegistration and TestToolSchemas
(replacing bare try/finally blocks for mcp.enable/disable).
